pyproject.toml
[project]
name = "skadatachallenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "21cmfast>=4.1.0",
    "ipykernel>=7.2.0",
    "seaborn>=0.13.2",
    "tools21cm>=2.3.10",
    "torch>=2.11.0",
]
